kapua/requirements.txt
Django==2.0.7
django-cors-headers==2.4.0
django-filter==2.0.0
django-filters==0.2.1
django-tenants==2.0.0
django-treebeard==4.3
djangorestframework==3.8.2
Markdown==2.6.11
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
